Fixed a type-size bug

+++ kernel/h/common.h
@@ -12,10 +12,10 @@ typedef char s8;
 typedef unsigned char u8;
 typedef short s16;
 typedef unsigned short u16;
-typedef int s32;
-typedef unsigned int u32;
-typedef long s64;
-typedef unsigned long u64;
+typedef long s32;
+typedef unsigned long u32;
+typedef long long s64;
+typedef unsigned long long u64;
 
 typedef u8 bool;
 typedef s8* string;
